start separating out cvxpygen tests
