+++ .github/workflows/cvxpygen.yml
@@ -27,7 +27,7 @@ jobs:
           uv pip install -e "." "cvxpy @ git+https://github.com/cvxpy/cvxpy.git"
           git clone --recurse-submodules https://github.com/cvxgrp/cvxpygen.git
           cd cvxpygen
-          uv pip install ".[dev]" "torch>=2.3" "jax==0.5.3"
+          uv pip install ".[dev]"
       - name: Run CVXPYgen tests
         run: |
-          uv run --no-sync pytest cvxpygen/tests/test_diff.py
+          uv run --no-sync pytest tests/test_cvxpygen.py
