+++ tests/test_cvxpygen.py
@@ -1,0 +1,171 @@
+# ===========================================================================
+# CVXPYgen integration — uses SolverInterface.from_parametric_functions()
+# ===========================================================================
+# Requires:  pip install cvxpygen
+#            cmake >= 3.5 (to compile the generated C extension)
+#
+# The three cpg functions generated by CVXPYgen are passed directly to
+# SolverInterface.from_parametric_functions().  All problem specification
+# (variables, parameters, cone dimensions) is owned by CvxpyLayer; the solver
+# just provides the compiled code.  Gradients flow directly from param.gradient
+# back to the input parameter tensors — no pseudoinverse needed.
+# ===========================================================================
+
+import os
+import subprocess
+import sys
+import pytest
+
+torch = pytest.importorskip("torch")
+import cvxpy as cp  # noqa: E402
+
+from cvxpylayers.interfaces.base import SolverInterface  # noqa: E402
+from cvxpylayers.torch import CvxpyLayer  # noqa: E402
+
+torch.set_default_dtype(torch.double)
+
+
+def _random_inputs(m, n, seed=0):
+    torch.manual_seed(seed)
+    A_t = torch.randn(m, n, requires_grad=True, dtype=torch.double)
+    b_t = torch.randn(m, requires_grad=True, dtype=torch.double)
+    return A_t, b_t
+
+
+def _make_problem():
+    """Simple LP: min 0.5*||Ax - b||_1  s.t. x >= 0."""
+    n, m = 2, 3
+    x = cp.Variable(n)
+    A = cp.Parameter((m, n))
+    b = cp.Parameter(m)
+    objective = cp.Minimize(0.5 * cp.pnorm(A @ x - b, p=1))
+    problem = cp.Problem(objective, [x >= 0])
+    assert problem.is_dpp()
+    return problem, A, b, x
+
+
+def _reference_layer(problem, A_param, b_param, x_var):
+    return CvxpyLayer(problem, parameters=[A_param, b_param], variables=[x_var])
+
+
+def _cmake_available() -> bool:
+    try:
+        subprocess.run(["cmake", "--version"], capture_output=True, check=True)
+        return True
+    except (FileNotFoundError, subprocess.CalledProcessError):
+        return False
+
+
+pytestmark_cmake = pytest.mark.skipif(
+    not _cmake_available(), reason="cmake required to compile CVXPYgen C code"
+)
+
+
+@pytest.fixture
+def problem_and_ref():
+    problem, A_param, b_param, x_var = _make_problem()
+    ref = _reference_layer(problem, A_param, b_param, x_var)
+    return problem, A_param, b_param, x_var, ref
+
+
+def _generate_cpg_code(problem, tmp_path, name, solver="OSQP"):
+    """Generate CVXPYgen C code in tmp_path/name, compile it, and return the module.
+
+    CVXPYgen internally calls importlib.import_module(f'{code_dir}.cpg_solver'),
+    so code_dir must be a plain module name, not an absolute path.  We chdir
+    into tmp_path and pass only the short name so that both the file-system
+    writes and the import resolve correctly.
+    """
+    import importlib
+    cpg = pytest.importorskip("cvxpygen.cpg")
+    parent_str = str(tmp_path)
+    if parent_str not in sys.path:
+        sys.path.insert(0, parent_str)
+    old_cwd = os.getcwd()
+    try:
+        os.chdir(parent_str)
+        cpg.generate_code(problem, code_dir=name, solver=solver, gradient=True, prefix=name)
+    finally:
+        os.chdir(old_cwd)
+    return importlib.import_module(f"{name}.cpg_solver")
+
+
+@pytestmark_cmake
+def test_cvxpygen_forward(problem_and_ref, tmp_path):
+    """from_parametric_functions forward solution matches the reference diffcp solution."""
+    problem, A_param, b_param, x_var, ref = problem_and_ref
+    n, m = 2, 3
+    cpg_mod = _generate_cpg_code(problem, tmp_path, "cpg_fwd", solver="OSQP")
+
+    layer = CvxpyLayer(
+        problem, parameters=[A_param, b_param], variables=[x_var],
+        solver=SolverInterface.from_parametric_functions(
+            solve               = cpg_mod.cpg_solve,
+            solve_and_gradient  = cpg_mod.cpg_solve_and_gradient_info,
+            gradient            = cpg_mod.cpg_gradient,
+        ),
+    )
+
+    A_t, b_t = _random_inputs(m, n)
+    (x_ref,) = ref(A_t, b_t)
+    (x_custom,) = layer(A_t, b_t)
+
+    assert x_custom.shape == x_ref.shape
+    assert torch.allclose(x_custom, x_ref, atol=1e-4)
+
+
+@pytestmark_cmake
+def test_cvxpygen_backward(problem_and_ref, tmp_path):
+    """from_parametric_functions backward gradients match the reference diffcp gradients."""
+    problem, A_param, b_param, x_var, ref = problem_and_ref
+    n, m = 2, 3
+    cpg_mod = _generate_cpg_code(problem, tmp_path, "cpg_bwd", solver="OSQP")
+
+    layer = CvxpyLayer(
+        problem, parameters=[A_param, b_param], variables=[x_var],
+        solver=SolverInterface.from_parametric_functions(
+            solve               = cpg_mod.cpg_solve,
+            solve_and_gradient  = cpg_mod.cpg_solve_and_gradient_info,
+            gradient            = cpg_mod.cpg_gradient,
+        ),
+    )
+
+    # Compute reference gradients
+    A_t, b_t = _random_inputs(m, n)
+    (x_ref,) = ref(A_t, b_t)
+    x_ref.sum().backward()
+    dA_ref, db_ref = A_t.grad.clone(), b_t.grad.clone()
+
+    # Compute CVXPYgen gradients at the same inputs
+    A_t2 = A_t.clone().detach().requires_grad_(True)
+    b_t2 = b_t.clone().detach().requires_grad_(True)
+    (x_custom,) = layer(A_t2, b_t2)
+    x_custom.sum().backward()
+
+    assert torch.allclose(A_t2.grad, dA_ref, atol=1e-3)
+    assert torch.allclose(b_t2.grad, db_ref, atol=1e-3)
+
+
+@pytestmark_cmake
+def test_cvxpygen_updated_params(problem_and_ref, tmp_path):
+    """updated_params passed via solver_args flows through to the cpg functions."""
+    problem, A_param, b_param, x_var, ref = problem_and_ref
+    n, m = 2, 3
+    cpg_mod = _generate_cpg_code(problem, tmp_path, "cpg_upd", solver="OSQP")
+
+    layer = CvxpyLayer(
+        problem, parameters=[A_param, b_param], variables=[x_var],
+        solver=SolverInterface.from_parametric_functions(
+            solve               = cpg_mod.cpg_solve,
+            solve_and_gradient  = cpg_mod.cpg_solve_and_gradient_info,
+            gradient            = cpg_mod.cpg_gradient,
+        ),
+    )
+
+    A_t, b_t = _random_inputs(m, n)
+    (x_ref,) = ref(A_t, b_t)
+    # updated_params=None signals that all parameters were updated (same as omitting it).
+    (x_custom,) = layer(A_t, b_t, solver_args={"updated_params": None})
+
+    assert x_custom.shape == x_ref.shape
+    assert torch.allclose(x_custom, x_ref, atol=1e-4)
